Update MangaGeko.kt
What is done ?
Added extractcoverurl , accessing to high quality images from the site and also fallback if none found. this gets rid of the issue in mangageko where some mangas don't have covers due to kotatsu accessing cover page from the divisions first image URL method which sometimes may not consist of the URL for the image of manga and gives a null feedback.

Con:
this works only for mgeko.cc web as that site had "div.max-image img" containing image , if there is error fetching image later on should check this div to change the img selector details.