Update MangasNoSekai.kt
what changes ?
Added variable for pagination and spun the code around it to get the chapters on app which were not shown , the page of said available chapters on kotatsu are all available..

Why the changes ? (kinda imp ig)

the site has gives links to the recent chapter pages on the 1st page of site and on further scooping i found link for the very 1st chapter too so my guess it it made a automated link with a refence to how link structure would be so it can be accessed and hosted by web. to access it i have added a loop for pages which will keep running until the chapter link comes out as nada i.e. not found and thus maybe the issue could be solved . 

note :
haven't checked its working so yeah do so